move helpers to duration file + convert duration to ts
