+++ shell/utils/time.ts
@@ -143,58 +143,3 @@ export const dateTimeFormat = (value: string | undefined, store: Store<any>): st
 
   return day(value).format(format);
 };
-
-const DURATION_UNITS = [
-  { divisor: 86400, suffix: 'd' },
-  { divisor: 3600, suffix: 'h' },
-  { divisor: 60, suffix: 'm' },
-  { divisor: 1, suffix: 's' },
-];
-
-/**
- * Decomposes a total number of seconds into the largest whole time unit.
- * For example, 7200 returns { value: 2, unit: 3600 } (2 hours).
- * @param seconds - Total seconds to decompose
- * @returns An object with `value` (the count) and `unit` (the multiplier in seconds: 86400, 3600, 60, or 1)
- */
-export function secondsToLargestUnit(seconds: number): { value: number, unit: number } {
-  if (seconds <= 0) {
-    return { value: seconds, unit: 1 };
-  }
-
-  if (seconds % 86400 === 0) {
-    return { value: seconds / 86400, unit: 86400 };
-  }
-  if (seconds % 3600 === 0) {
-    return { value: seconds / 3600, unit: 3600 };
-  }
-  if (seconds % 60 === 0) {
-    return { value: seconds / 60, unit: 60 };
-  }
-
-  return { value: seconds, unit: 1 };
-}
-
-/**
- * Formats a duration in seconds into a human-readable string (e.g. "1d 3h 46m 40s").
- * Zero-value components are omitted. Returns "0s" for non-positive values.
- * @param seconds - Duration in seconds
- */
-export function formatDuration(seconds: number): string {
-  if (seconds <= 0) {
-    return '0s';
-  }
-
-  let remaining = seconds;
-
-  return DURATION_UNITS
-    .map(({ divisor, suffix }) => {
-      const count = Math.floor(remaining / divisor);
-
-      remaining %= divisor;
-
-      return count > 0 ? `${ count }${ suffix }` : '';
-    })
-    .filter(Boolean)
-    .join(' ');
-}
